% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_to_pmml}
\alias{export_to_pmml}
\title{Export Model to PMML}
\usage{
export_to_pmml(self, file_path = "model.pmml")
}
\arguments{
\item{self}{The MNLMIXTE object.}

\item{file_path}{Character. File path for the PMML model.}
}
\value{
None. Saves the PMML model.
}
\description{
Export Model to PMML
}
