% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Train the Multinomial Logistic Regression Model}
\usage{
fit_model(self, X, y, variable_selection = FALSE, use_parallel = FALSE)
}
\arguments{
\item{self}{An instance of the MNLMIXTE class.}

\item{X}{A matrix or data frame of features.}

\item{y}{A factor vector of target labels.}

\item{variable_selection}{Logical. If \code{TRUE}, performs variable selection. Default: FALSE.}

\item{use_parallel}{Logical. If \code{TRUE}, enables parallel computation. Default: FALSE.}
}
\value{
None. Updates the model's coefficients and loss history in the MNLMIXTE instance.
}
\description{
This function trains the multinomial logistic regression model on the provided dataset.
}
