% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot Variable Importance}
\usage{
plot_feature_importance(self)
}
\value{
None. Displays a bar plot.
}
\description{
Visualizes the importance of variables using a bar plot.
}
\examples{
\dontrun{
#' # Plot feature importance
model$plot_importance()
}
}
