% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_training_loss}
\alias{plot_training_loss}
\title{Plot Training Loss}
\usage{
plot_training_loss(self)
}
\value{
None. Displays a plot of the loss function.
}
\description{
Visualizes the evolution of the loss function during training.
}
\examples{
\dontrun{
# Plot the loss history
model$plot_loss()
}
}
