% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_labels}
\alias{predict_labels}
\title{Predict Class Labels}
\usage{
predict_labels(self, X)
}
\arguments{
\item{self}{The \code{MNLMIXTE} R6 object.}

\item{X}{A matrix or data frame of features.}
}
\value{
A factor vector of predicted class labels.
}
\description{
Predicts the class labels based on the input data and the model coefficients.
}
