% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_probabilities}
\alias{predict_probabilities}
\title{Predict Class Probabilities}
\usage{
predict_probabilities(self, X)
}
\arguments{
\item{self}{The \code{MNLMIXTE} R6 object.}

\item{X}{A matrix or data frame of features.}
}
\value{
A matrix of predicted probabilities, where each column corresponds to a class.
}
\description{
Predicts the probabilities for each class based on the input data and the model coefficients.
}
