% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data}
\usage{
preprocess_data(X)
}
\arguments{
\item{X}{A data frame or matrix of features.}
}
\value{
A preprocessed matrix.
}
\description{
Preprocess Data
}
