% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softmax}
\alias{softmax}
\title{Softmax Function}
\usage{
softmax(z)
}
\arguments{
\item{z}{A numeric matrix of scores.}
}
\value{
A matrix of probabilities after applying the softmax transformation.
}
\description{
Applies the softmax function to a matrix of scores.
}
